add style popup

+++ src/app/status/status.component.ts
@@ -5,10 +5,11 @@ import {CommentService} from '../service/comment.service';
 import {IPost} from '../model/IPost';
 import {IUser} from '../model/iuser';
 import {IComment} from '../model/IComment';
-import {ActivatedRoute} from '@angular/router';
+import {ActivatedRoute, Router} from '@angular/router';
 import {LikePostService} from '../service/like-post.service';
 import {ILikePost} from '../model/ILikePost';
 import {TokenStorageService} from '../service/signin-signup/token-storage.service';
+import swal from 'sweetalert';
 
 @Component({
   selector: 'app-status',
@@ -22,7 +23,8 @@ export class StatusComponent implements OnInit {
               private commentService: CommentService,
               private likePostService: LikePostService,
               private tokenStorage: TokenStorageService,
-              private actRoute: ActivatedRoute) { }
+              private actRoute: ActivatedRoute,
+              private router: Router) { }
 
   ngOnInit(): void {
     this.showPost();
@@ -124,21 +126,38 @@ export class StatusComponent implements OnInit {
   }
 
   deletePost(postId: any) {
-    this.commentService.getCommentByPostId(postId).subscribe(
-      commentList => {
-        let comments = <IComment[]> commentList;
-        for (let i = 0; i < comments.length; i++) {
-          this.commentService.deleteComment(comments[i].commentId).subscribe(
-            res => console.log("comment deleted")
+    swal({
+      title: "Are you sure?",
+      text: "Are you sure that you want to delete this post?",
+      icon: "warning",
+      dangerMode: true,
+    })
+      .then(willDelete => {
+        if (willDelete) {
+          this.commentService.getCommentByPostId(postId).subscribe(
+            commentList => {
+              let comments = <IComment[]> commentList;
+              for (let i = 0; i < comments.length; i++) {
+                this.commentService.deleteComment(comments[i].commentId).subscribe(
+                  res => console.log("comment deleted")
+                )
+              }
+            }
+          )
+          this.postService.deletePost(postId).subscribe(
+            res => {
+              swal({
+                icon: "success",
+                title: "Your post has been deleted!"
+              });
+              if (this.actRoute.snapshot.params.id != null) {
+                this.router.navigate(['/home']);
+              }
+            }
           )
         }
       }
-    )
-    this.postService.deletePost(postId).subscribe(
-      res => {
-        window.alert("Post deleted");
-      }
-    )
+    );
   }
 
   addNewComment(value) {
